requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langchain_groq
groq
langchain-text-splitters
faiss-cpu
pymupdf
sentence-transformers
streamlit
python-dotenv
